requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# app.py: 1
# core\auth.py: 4
# core\places.py: 4
# migrations\env.py: 20
# views\test.py: 1
Flask == 0.12
gunicorn==19.3.0
psycopg2

# app.py: 2
Flask_Cors == 3.0.3

Flask_Mail == 0.9.1

# app.py: 4
# manage.py: 2
Flask_Migrate == 2.0.2

# models.py: 1
Flask_SQLAlchemy == 2.1

# app.py: 3
# manage.py: 1
Flask_Script == 2.0.5

# core\auth.py: 7
PyJWT == 1.5.3

# core\auth.py: 3
# core\places.py: 1
# migrations\env.py: 3
# migrations\versions\3fc5958ddcb2_.py: 9
# migrations\versions\573bb0068a15_.py: 9
# migrations\versions\b8bc7bb95412_.py: 9
# migrations\versions\dcac06755a95_.py: 9
# migrations\versions\e5227e6c07d4_.py: 9
# migrations\versions\e9b0becfe76e_.py: 9
SQLAlchemy == 1.1.2

# core\auth.py: 1
Werkzeug == 0.11.11

# migrations\env.py: 2
# migrations\versions\3fc5958ddcb2_.py: 8
# migrations\versions\573bb0068a15_.py: 8
# migrations\versions\b8bc7bb95412_.py: 8
# migrations\versions\dcac06755a95_.py: 8
# migrations\versions\e5227e6c07d4_.py: 8
# migrations\versions\e9b0becfe76e_.py: 8
alembic == 0.8.10

# api\__init__.py: 1
# api\auth.py: 1
# api\places.py: 1
flask_restplus == 0.10.1

# core\places.py: 2
# dev_utils.py: 1
python_google_places == 1.4.1

# core\places.py: 3
# dev_utils.py: 2
yelpapi == 2.0.1
